Module ori_core::canvas

Canvas module.

Structs

BorderRadius

Radi of the corners on a rounded rectangle.

BorderWidth

The border width of a rounded rectangle.

Canvas

A canvas that can be drawn on.

CanvasDiff

A canvas that can be drawn on.

Color

A color with red, green, blue and alpha components.

Curve

A bezier curve.

CurveIter

An iterator over the segments of a curve.

DisplayHex

A type for displaying a color as a hex string.

Mask

A mask that can be used to clip a layer.

Paint

A paint that can be used to fill or stroke a shape.

Pattern

A pattern that can be used to fill a shape.

Stroke

Properties of a stroke.

Enums

AntiAlias

Ways to anti-alias a shape.

BlendMode

Ways to blend two colors.

CurveSegment

A segment of a curve.

CurveVerb

A verb that describes the type of curve.

FillRule

Rule determining if a point is inside a shape.

LineCap

Ways to draw the end of a line.

LineJoin

Ways to join two lines.

Primitive

A primitive that can be drawn on a canvas.

Shader

Ways to fill a shape.

Functions

hex

Create a new color, with the given hex string.

hsl

Create a new color, with the given hue, saturation and lightness components.

hsla

Create a new color, with the given hue, saturation, lightness and alpha components.

hsv

Create a new color, with the given hue, saturation and value components.

hsva

Create a new color, with the given hue, saturation, value and alpha components.

okhsl

Create a new color, with the given hue, saturation and lightness components.

okhsla

Create a new color, with the given hue, saturation, lightness and alpha components.

okhsv

Create a new color, with the given hue, saturation and value components.

okhsva

Create a new color, with the given hue, saturation, value and alpha components.

oklab

Create a new color, with the given lightness, a and b components.

oklaba

Create a new color, with the given lightness, a, b and alpha components.

oklch

Create a new color, with the given lightness, chroma and hue components.

oklcha

Create a new color, with the given lightness, chroma, hue and alpha components.

rgb

Create a new color, with the given red, green and blue components.

rgba

Create a new color, with the given red, green, blue and alpha components.