ori_core::context
Struct BaseCx

pub struct BaseCx<'a> { /* private fields */ }

A base context that is shared between all other contexts.

Implementations

impl<'a> BaseCx<'a>

pub fn new(contexts: &'a mut Contexts, proxy: &'a mut CommandProxy) -> Self

Create a new base context.

pub fn fonts(&self) -> &dyn Fonts

Get the Fonts.

pub fn fonts_mut(&mut self) -> &mut dyn Fonts

Get the Fonts as a mutable reference.

pub fn clipboard(&mut self) -> &mut Clipboard

Get the Clipboard.

pub fn proxy(&self) -> CommandProxy

Get the CommandProxy.

pub fn cmd<T: Any + Send>(&mut self, command: T)

Emit a command.

pub fn spawn_async(&mut self, future: impl Future<Output = ()> + Send + 'static)

Spawn a future.

pub fn cmd_async<T: Any + Send>( &self, future: impl Future<Output = T> + Send + 'static, )

Spawn a future sending a command when it completes.

pub fn contexts(&self) -> &Contexts

Get a reference to the Contexts.

pub fn contexts_mut(&mut self) -> &mut Contexts

Get a mutable reference to the Contexts.

pub fn insert_context<T: Any>(&mut self, context: T) -> Option<T>

Insert a context.

pub fn contains_context<T: Any>(&self) -> bool

Check if a context is contained.

pub fn remove_context<T: Any>(&mut self) -> Option<T>

Remove a context.

pub fn get_context<T: Any>(&self) -> Option<&T>

Get a context.

pub fn get_context_mut<T: Any>(&mut self) -> Option<&mut T>

Get a mutable context.

pub fn context<T: Any>(&self) -> &T

Get a context.

Panics

If the context is not found.

pub fn context_mut<T: Any>(&mut self) -> &mut T

Get a mutable context.

Panics

If the context is not found.

pub fn context_or_default<T: Any + Default>(&mut self) -> &mut T

Get a context or insert a default.