ori_core::context
Struct BuildCx

pub struct BuildCx<'a, 'b> { /* private fields */ }

A context for building the view tree.

Implementations

impl<'a, 'b> BuildCx<'a, 'b>

pub fn new(base: &'a mut BaseCx<'b>, view_state: &'a mut ViewState) -> Self

Create a new build context.

pub fn child(&mut self) -> BuildCx<'_, 'b>

Create a child context.

impl BuildCx<'_, '_>

pub fn window(&self) -> &Window

Get the window.

pub fn window_mut(&mut self) -> &mut Window

Get the window mutably.

pub fn styles(&self) -> &Styles

Get the styles.

pub fn id(&self) -> ViewId

Get the id of the view.

pub fn is_hovered(&self) -> bool

Get whether the view is hovered.

pub fn is_focused(&self) -> bool

Get whether the view is focused.

pub fn is_active(&self) -> bool

Get whether the view is active.

pub fn is_focusable(&self) -> bool

Get whether the view is focusable.

pub fn has_hovered(&self) -> bool

Get whether a child view is hovered.

pub fn has_focused(&self) -> bool

Get whether a child view is focused.

pub fn has_active(&self) -> bool

Get whether a child view is active.

pub fn contains_property<T: 'static>(&self) -> bool

Check if the view has the property T.

pub fn insert_property<T: 'static>(&mut self, item: T)

Insert a property into the view.

pub fn remove_property<T: 'static>(&mut self) -> Option<T>

Remove a property from the view.

pub fn get_property<T: 'static>(&self) -> Option<&T>

Get the property T of the view.

pub fn get_property_mut<T: 'static>(&mut self) -> Option<&mut T>

Get the property T of the view mutably.

pub fn property_or_insert_with<T: 'static, F: FnOnce() -> T>( &mut self, f: F, ) -> &mut T

Get the property T of the view or insert it with a value.

pub fn property_or<T: 'static>(&mut self, item: T) -> &mut T

Get the property T of the view or insert it with a value.

pub fn property_or_default<T: 'static + Default>(&mut self) -> &mut T

Get the property T of the view or insert it with a default value.

pub fn class(&self) -> Option<&str>

Get the class of the view.

impl BuildCx<'_, '_>

pub fn propagate(&mut self, child: &mut ViewState)

Propagate the view state of a child view.

pub fn layout(&mut self)

Request a layout of the view tree.

pub fn draw(&mut self)

Request a draw of the view tree.

pub fn animate(&mut self)

Request an animation frame.

pub fn focus(&mut self)

Request focus for the view.

pub fn focus_next(&mut self)

Request the next focusable view to be focused.

pub fn focus_prev(&mut self)

Request the previous focusable view to be focused.

pub fn set_cursor(&mut self, cursor: Option<Cursor>)

Set the cursor of the view.

pub fn set_ime(&mut self, ime: Option<Ime>)

Set the ime of the view.

pub fn set_hovered(&mut self, hovered: bool) -> bool

Set whether the view is hovered.

Returns true if the hovered state changed.

pub fn set_focused(&mut self, focused: bool) -> bool

Set whether the view is focused.

Returns true if the focused state changed.

pub fn set_active(&mut self, active: bool) -> bool

Set whether the view is active.

Returns true if the active state changed.

pub fn set_focusable(&mut self, focusable: bool) -> bool

Set whether the view is focusable.

Returns true if the focusable state changed.

pub fn set_class(&mut self, class: impl Into<String>)

Set the class of the view.

This should be called before trying to get any style properties.