Module ori_core::debug

Debugging and profiling tools.

Structs

BuildItem

Emitted when a window builds it’s view tree.

DebugDraw

Debug information about how a debug tree was drawn.

DebugLayout

Debug information about how a debug tree was laid out.

DebugTree

A debug tree.

DrawItem

Emitted when a window draws it’s view tree.

EventItem

Emitted when a window handles an event.

History

A history of events that have occurred in the application.

LayoutItem

Emitted when a window lays out it’s view tree.

RebuildItem

Emitted when a window rebuilds it’s view tree.

Enums

HistoryItem

An item in the history of the application.

Functions

debug_ui

Wrap a view in a debug view.