Module ori_core::event

Events and event handling.

Structs

AnimationFrame

An animation frame, containing the time since the last frame.

CloseRequested

Event emitted when a window wants to close.

Event

An event that can be sent to a view.

KeyPressed

An event fired when a key is pressed.

KeyReleased

An event fired when a key is released.

Modifiers

Modifiers for keyboard and pointer events.

Pointer

A pointer.

PointerClicked

A pointer was clicked.

PointerId

A unique pointer id.

PointerLeft

A pointer left the window.

PointerMoved

A pointer was moved.

PointerPressed

A pointer button was pressed.

PointerReleased

A pointer button was released.

PointerScrolled

A pointer wheel was scrolled.

Quit

Command sent to quit the application.

WindowResized

Event emitted when a window is resized.

Enums

Code

A keyboard key.

PointerButton

A pointer button.

RequestFocus

Event emitted when a view should be focused.

SwitchFocus

Event emitted when focus should be switched.