ori_core
Module text

Text rendering and layouting.

Macros

include_font

Load a font from a file or directory.

Structs

FontAttributes

Attributes of a section of text.

FontWeight

A font weight.

GlyphCluster

A glyph cluster in a line of laid out text.

Paragraph

A paragraph of rich text, that can contain multiple segments with different FontAttributes.

TextLayoutLine

A line of text layout.

Enums

FontFamily

A font family, by default FontFamily::SansSerif.

FontSource

A source for a font.

FontStretch

A font stretch.

FontStyle

A font style.

TextAlign

Alignment of a section of text.

TextDirection

The direction of text.

TextWrap

Wrapping of a section of text.

Traits

Fonts

A trait for managing fonts and text layout.