Struct ori_core::view::PodSeq

pub struct PodSeq<V> { /* private fields */ }

Contents of a view, in a sequence.

This is useful for views that contain multiple pieces of content. See ViewSeq for more information.

See Pod for more information on when to use this.

Implementations

impl<V> PodSeq<V>

pub fn new(views: V) -> Self

Create a new PodSeq.

impl<V> PodSeq<V>

pub fn len<T>(&self) -> usize
where V: ViewSeq<T>,

The length of the sequence.

pub fn is_empty<T>(&self) -> bool
where V: ViewSeq<T>,

Whether the sequence is empty.

pub fn build<T>( &mut self, cx: &mut BuildCx<'_, '_>, data: &mut T ) -> SeqState<T, V>
where V: ViewSeq<T>,

Build the sequence state.

pub fn rebuild<T>( &mut self, state: &mut SeqState<T, V>, cx: &mut BuildCx<'_, '_>, data: &mut T, old: &Self )
where V: ViewSeq<T>,

Rebuild the sequence state.

pub fn rebuild_nth<T>( &mut self, n: usize, state: &mut SeqState<T, V>, cx: &mut RebuildCx<'_, '_>, data: &mut T, old: &Self )
where V: ViewSeq<T>,

Rebuild the nth view.

pub fn event_nth<T>( &mut self, n: usize, state: &mut SeqState<T, V>, cx: &mut EventCx<'_, '_>, data: &mut T, event: &Event )
where V: ViewSeq<T>,

Handle an event for the nth view.

pub fn layout_nth<T>( &mut self, n: usize, state: &mut SeqState<T, V>, cx: &mut LayoutCx<'_, '_>, data: &mut T, space: Space ) -> Size
where V: ViewSeq<T>,

Layout the nth view.

pub fn draw_nth<T>( &mut self, n: usize, state: &mut SeqState<T, V>, cx: &mut DrawCx<'_, '_>, data: &mut T, canvas: &mut Canvas<'_> )
where V: ViewSeq<T>,

Draw the nth view.