ori_core
Module views

The builtin views in Ori.

Macros

format_text

Create a formatted Text.

hstack

Create a horizontal Stack.

hwrap

Create a horizontal Wrap.

text

Create a formatted Text.

vstack

Create a vertical Stack.

vwrap

Create a vertical Wrap.

zstack

Create a new ZStack view.

Structs

Aligned

A view that aligns its content.

Animate

A view that animates.

Aspect

A view that lays out its content with a fixed aspect ratio.

BuildHandler

A view that hooks into the build cycle.

Builder

A view that is built from a closure.

Button

A button.

ButtonStyle

The style of a button.

Checkbox

A checkbox.

CheckboxStyle

The style of a checkbox.

Clickable

A click handler.

Collapsing

A collapsing view.

CollapsingStyle

The style of a collapsing view.

ColorPicker

A color picker.

ColorPickerStyle

The style of a color picker.

Constrain

A view that constrains its content to a given space.

Container

A container view.

ContainerStyle

The style of a container.

Decorate

A view for creating custom window decorations.

DrawHandler

A view that hooks into the draw cycle.

EventHandler

A view that handles events.

Flex

The flex value of a view.

Flexible

A flexible view.

Focus

A view that focuses on a part of the data.

Layout

A view for creating views based on the available space.

Memo

A view that only builds the inner view when certain data changes.

Opaque

A view that bridges the gap between impl View and impl View<T>.

Pad

A view that adds padding to its content.

Painter

A view that draws something.

Popup

A view that can popup.

RebuildHandler

A view that handles rebuilds.

Scroll

A scrollable view.

ScrollStyle

The style of a scroll view.

Slider

A slider.

SliderStyle

The style of a Slider.

Stack

A view that stacks it’s content in a line.

StackStyle

A style for a Stack.

Suspense

A view that suspends rendering while a future is pending.

Text

A view that displays text.

TextInput

A text input.

TextInputStyle

The style of a TextInput.

TextStyle

The style of a Text.

Tooltip

A view that displays some text when the content is hovered.

TooltipStyle

The style of a Tooltip.

Transform

A view that transforms its content.

Trigger

A view that creates a trigger around the content.

WithState

A view that stores some additional data.

Wrap

A view that lays out it’s content in a line wrapping if it doesn’t fit.

WrapStyle

The style of a Wrap.

ZStack

A view that overlays its content on top of each other.

Enums

ClickEvent

A click event.

Functions

align

Create a new Aligned view.

animate

Create a new Animate.

aspect

Create a new Aspect view.

background

Create a new Container with background.

bottom

Create a new Aligned view that aligns its content to the bottom.

bottom_left

Create a new Aligned view that aligns its content to the bottom left.

bottom_right

Create a new Aligned view that aligns its content to the bottom right.

build

Build a view from a closure.

button

Create a new Button.

center

Create a new Aligned view that aligns its content to the center.

checkbox

Create a new Checkbox.

circle

Create a new Painter view that draws a circle.

collapsing

Create a new Collapsing.

color_picker

Create a new ColorPicker.

constrain

Create a new Constrained view, constraining its content to a space.

container

Create a new Container.

decorate

A view for creating custom window decorations.

ellipse

Create a new Painter view that draws an ellipse.

expand

Create a new expanded Flexible view.

flex

Create a new Flexible view.

focus

Create a new Focus.

height

Create a new Constrained view, constraining its content to a height.

hscroll

Create a new horizontal Scroll.

hstack

Create a horizontal Stack.

hstack_any

Create a horizontal Stack, with dynamic content.

hstack_vec

Create a horizontal Stack, with vector content.

hwrap

Create a horizontal Wrap.

hwrap_any

Create a horizontal Wrap, with dynamic content.

hwrap_vec

Create a horizontal Wrap, with a vector of content.

layout

A view for creating views based on the available space.

left

Create a new Aligned view that aligns its content to the left.

max_height

Create a new Constrained view, constraining its content to a maximum height.

max_size

Create a new Constrained view, constraining its content to a maximum size.

max_width

Create a new Constrained view, constraining its content to a maximum width.

memo

Create a new Memo.

min_height

Create a new Constrained view, constraining its content to a minimum height.

min_size

Create a new Constrained view, constraining its content to a minimum size.

min_width

Create a new Constrained view, constraining its content to a minimum width.

on_build

Create a new BuildHandler.

on_click

Create a new Clickable, that calls on_click when clicked.

on_draw

Create a new DrawHandler.

on_event

Create a new EventHandler, with an after callback.

on_event_before

Create a new EventHandler, with a before callback.

on_press

Create a new Clickable, that calls on_press when pressed.

on_rebuild

Create a new RebuildHandler.

on_release

Create a new Clickable, that calls on_release when released.

opaque

A view that bridges the gap between impl View and impl View<T>.

pad

Create a new Pad view.

pad_bottom

Create a new Pad view adding padding to the bottom.

pad_left

Create a new Pad view adding padding to the left.

pad_right

Create a new Pad view adding padding to the right.

pad_top

Create a new Pad view adding padding to the top.

painter

Create a new Painter view.

popup

Create a new popup view.

rect

Create a new Painter view that draws a rectangle.

right

Create a new Aligned view that aligns its content to the right.

rotate

Create a new Transform view that rotates its content.

rotate_degrees

Create a new Transform view that rotates its content in degrees.

scale

Create a new Transform view that scales its content.

size

Create a new Constrained view, cosntraining its content to a size.

slider

Create a new Slider.

suspense

Create a new Suspense view.

text

Create a new Text.

text_input

Create a new TextInput.

tooltip

Create a new Tooltip view.

top

Create a new Aligned view that aligns its content to the top.

top_left

Create a new Aligned view that aligns its content to the top left.

top_right

Create a new Aligned view that aligns its content to the top right.

transform

Create a new Transform view.

transition

Animate a transition.

transition_active

Animate a view when active changes.

transition_focused

Animate a view when focused changes.

transition_hovered

Animate a view when hovered changes.

translate

Create a new Transform view that translates its content.

trigger

Create a new Trigger view.

vscroll

Create a new vertical Scroll.

vstack

Create a vertical Stack.

vstack_any

Create a vertical Stack, with dynamic content.

vstack_vec

Create a vertical Stack, with vector content.

vwrap

Create a vertical Wrap.

vwrap_any

Create a vertical Wrap, with dynamic content.

vwrap_vec

Create a vertical Wrap, with a vector of content.

width

Create a new Constrained view, constraining its content to a width.

with_data

Create a new WithState that replaces the data with the state.

with_data_default

Create a new WithState that replaces the data with the state using S::default().

with_state

Create a new WithState.

with_state_default

Create a new WithState using S::default().

without_data

Create a new view unwrapping some data from the state.

without_state

Create a new view unwrapping some state from the data.

zstack

Create a new ZStack view.

Type Aliases

Lens

A lens used by Focus.