ori_core::views
Struct Stack

pub struct Stack<V> {
    pub content: PodSeq<V>,
    pub axis: Axis,
    pub justify: Option<Justify>,
    pub align: Option<Align>,
    pub gap: Option<f32>,
}

A view that stacks it’s content in a line.

Fields
content: PodSeq<V>

The content of the stack.

axis: Axis

The axis of the stack.

justify: Option<Justify>

How to justify the content along the main axis.

align: Option<Align>

How to align the content along the cross axis, within each line.

gap: Option<f32>

The gap between children.

Implementations

impl<V> Stack<V>

pub fn axis(self, axis: impl Into<Axis>) -> Self

Set self.axis.

The axis of the stack.

pub fn justify(self, justify: impl Into<Justify>) -> Self

Set self.justify.

How to justify the content along the main axis.

pub fn align(self, align: impl Into<Align>) -> Self

Set self.align.

How to align the content along the cross axis, within each line.

pub fn gap(self, gap: impl Into<f32>) -> Self

Set self.gap.

The gap between children.

impl<V> Stack<V>

pub fn new(axis: Axis, content: V) -> Self

Create a new Stack.

pub fn horizontal(content: V) -> Self

Create a new horizontal Stack.

pub fn vertical(content: V) -> Self

Create a new vertical Stack.

impl<T> Stack<Vec<T>>

pub fn vec(axis: Axis) -> Self

Create a new Stack, with vector content.

pub fn vec_horizontal() -> Self

Create a new horizontal Stack, with vector content.

pub fn vec_vertical() -> Self

Create a new vertical Stack, with vector content.

pub fn push(&mut self, view: T)

Push a view to the stack.

pub fn with(self, view: T) -> Self

Push a view to the stack.

pub fn is_empty(&self) -> bool

Get whether the stack is empty.

pub fn len(&self) -> usize

Get the number of views in the stack.

impl<T> Stack<Vec<Box<dyn AnyView<T> + '_>>>

pub fn any(axis: Axis) -> Self

Create a new Stack, with dynamic content.

pub fn any_horizontal() -> Self

Create a new horizontal Stack, with dynamic content.

pub fn any_vertical() -> Self

Create a new vertical Stack, with dynamic content.