Trait ori_core::window::RawWindow

pub trait RawWindow: Any {

    // Required methods
    fn title(&self) -> String;
    fn set_title(&mut self, title: &str);
    fn set_icon(&mut self, icon: Option<&Image>);
    fn size(&self) -> (u32, u32);
    fn set_size(&mut self, width: u32, height: u32);
    fn resizable(&self) -> bool;
    fn set_resizable(&mut self, resizable: bool);
    fn decorated(&self) -> bool;
    fn set_decorated(&mut self, decorated: bool);
    fn scale_factor(&self) -> f32;
    fn minimized(&self) -> bool;
    fn set_minimized(&mut self, minimized: bool);
    fn maximized(&self) -> bool;
    fn set_maximized(&mut self, maximized: bool);
    fn visible(&self) -> bool;
    fn set_visible(&mut self, visible: bool);
    fn set_cursor(&mut self, cursor: Cursor);
    fn color(&self) -> Option<Color>;
    fn set_color(&mut self, color: Option<Color>);
    fn set_soft_input(&mut self, visible: bool);
    fn drag(&mut self);
    fn request_draw(&mut self);

}

A wrapper around a raw window.

Required Methods

fn title(&self) -> String

Get the title of the window.

fn set_title(&mut self, title: &str)

Set the title of the window.

fn set_icon(&mut self, icon: Option<&Image>)

Set the icon of the window.

fn size(&self) -> (u32, u32)

Get the size of the window.

fn set_size(&mut self, width: u32, height: u32)

Set the size of the window.

fn resizable(&self) -> bool

Get whether the window is resizable.

fn set_resizable(&mut self, resizable: bool)

Set whether the window is resizable.

fn decorated(&self) -> bool

Get whether the window is decorated.

fn set_decorated(&mut self, decorated: bool)

Set whether the window is decorated.

fn scale_factor(&self) -> f32

Get the scale factor of the window.

fn minimized(&self) -> bool

Get whether the window is minimized.

fn set_minimized(&mut self, minimized: bool)

Set whether the window is minimized.

fn maximized(&self) -> bool

Get whether the window is maximized.

fn set_maximized(&mut self, maximized: bool)

Set whether the window is maximized.

fn visible(&self) -> bool

Get whether the window is visible.

fn set_visible(&mut self, visible: bool)

Set whether the window is visible.

fn set_cursor(&mut self, cursor: Cursor)

Set the cursor of the window.

fn color(&self) -> Option<Color>

Get the background of the window.

fn set_color(&mut self, color: Option<Color>)

Set the background of the window.

fn set_soft_input(&mut self, visible: bool)

Get whether soft input is enabled.

fn drag(&mut self)

Drag the window.

fn request_draw(&mut self)

Get whether the window is focused.

Implementations

impl dyn RawWindow

pub fn is<T: RawWindow>(&self) -> bool

Check if the window is of a specific type.

pub fn downcast_ref<T: RawWindow>(&self) -> Option<&T>

Try to downcast the window to a specific type.