Struct ori_core::context::EventCx

pub struct EventCx<'a, 'b> { /* private fields */ }

A context for handling events.

Implementations

impl<'a, 'b> EventCx<'a, 'b>

pub fn new( base: &'a mut BaseCx<'b>, view_state: &'a mut ViewState, rebuild: &'a mut bool, ) -> Self

Create a new event context.

pub fn child(&mut self) -> EventCx<'_, 'b>

Create a child context.

pub fn as_build_cx(&mut self) -> BuildCx<'_, 'b>

Get a build context.

pub fn as_rebuild_cx(&mut self) -> RebuildCx<'_, 'b>

Get a rebuild context.

pub fn size(&self) -> Size

Get the size of the view.

pub fn rect(&self) -> Rect

Get the rect of the view in local space.

pub fn transform(&self) -> Affine

Get the transform of the view.

pub fn local(&self, point: Point) -> Point

Transform a point from global space to local space.

pub fn rebuild(&mut self)

Request a rebuild of the view tree.

pub fn layout(&mut self)

Request a layout of the view tree.

pub fn draw(&mut self)

Request a draw of the view tree.

pub fn animate(&mut self)

Request an animation frame.

pub fn set_cursor(&mut self, cursor: Option<Cursor>)

Set the cursor of the view.

pub fn set_ime(&mut self, ime: Option<Ime>)

Set the ime of the view.

pub fn set_hovered(&mut self, hovered: bool) -> bool

Set whether the view is hovered.

Returns true if the hovered state changed.

pub fn set_focused(&mut self, focused: bool) -> bool

Set whether the view is focused.

Returns true if the focused state changed.

pub fn set_active(&mut self, active: bool) -> bool

Set whether the view is active.

Returns true if the active state changed.

pub fn was_hovered(&self) -> bool

Get whether the view was hovered last call.

pub fn was_focused(&self) -> bool

Get whether the view was focused last call.

pub fn was_active(&self) -> bool

Get whether the view was active last call.

pub fn had_hovered(&self) -> bool

Get whether a child view was hovered last call.

pub fn had_focused(&self) -> bool

Get whether a child view was focused last call.

pub fn had_active(&self) -> bool

Get whether a child view was active last call.

pub fn hovered_changed(&self) -> bool

Get whether the view’s hovered state changed.

pub fn focused_changed(&self) -> bool

Get whether the view’s focused state changed.

pub fn active_changed(&self) -> bool

Get whether the view’s active state changed.

pub fn has_hovered_changed(&self) -> bool

Get whether a child view’s hovered state changed.

pub fn has_focused_changed(&self) -> bool

Get whether a child view’s focused state changed.

pub fn has_active_changed(&self) -> bool

Get whether a child view’s active state changed.

impl EventCx<'_, '_>

pub fn window(&self) -> &Window

Get the window.

pub fn window_mut(&mut self) -> &mut Window

Get the window mutably.

pub fn styles(&self) -> &Styles

Get the styles.

pub fn id(&self) -> ViewId

Get the id of the view.

pub fn is_hovered(&self) -> bool

Get whether the view is hovered.

pub fn is_focused(&self) -> bool

Get whether the view is focused.

pub fn is_active(&self) -> bool

Get whether the view is active.

pub fn has_hovered(&self) -> bool

Get whether a child view is hovered.

pub fn has_focused(&self) -> bool

Get whether a child view is focused.

pub fn has_active(&self) -> bool

Get whether a child view is active.

pub fn contains_property<T: 'static>(&self) -> bool

Check if the view has the property T.

pub fn insert_property<T: 'static>(&mut self, item: T)

Insert a property into the view.

pub fn remove_property<T: 'static>(&mut self) -> Option<T>

Remove a property from the view.

pub fn get_property<T: 'static>(&self) -> Option<&T>

Get the property T of the view.

pub fn get_property_mut<T: 'static>(&mut self) -> Option<&mut T>

Get the property T of the view mutably.

pub fn property_or_insert_with<T: 'static, F: FnOnce() -> T>( &mut self, f: F, ) -> &mut T

Get the property T of the view or insert it with a value.

pub fn property_or<T: 'static>(&mut self, item: T) -> &mut T

Get the property T of the view or insert it with a value.

pub fn property_or_default<T: 'static + Default>(&mut self) -> &mut T

Get the property T of the view or insert it with a default value.