Module ori_core::layout

Layout of Views.

Structs

Affine

An affine transformation in 2 dimensional space.

Alignment

Alignment of content inside a container.

JustifyIterator

An iterator over the positions of items in a stack container.

Matrix

A 2x2 matrix.

Padding

A padding of a rectangle.

Point

A point in 2D space.

Rect

A rectangle defined by its minimum and maximum points.

Size

A 2 dimensional size.

Space

Space available to lay out a view.

Vector

A 2D vector.

Enums

Align

The alignment of items along the cross axis.

Axis

An axis is a direction in which a layout is applied.

Justify

The justify content of a stack container.

Constants

FILLDeprecation planned

Infinity (∞). Use f32::INFINITY instead.