Struct ori_core::layout::Vector

#[repr(C)]
pub struct Vector {
    pub x: f32,
    pub y: f32,
}

A 2D vector.

Fields
x: f32

The x coordinate.

y: f32

The y coordinate.

Implementations

impl Vector

pub const ZERO: Self = _

The zero vector.

pub const ONE: Self = _

The one vector.

pub const X: Self = _

The unit x vector.

pub const Y: Self = _

The unit y vector.

pub const NEG_X: Self = _

The negative unit x vector.

pub const NEG_Y: Self = _

The negative unit y vector.

pub const fn new(x: f32, y: f32) -> Self

Create a new vector.

pub const fn all(value: f32) -> Self

Create a new vector with the same x and y.

pub fn from_angle(angle: f32) -> Self

Create a new vector from an angle.

pub fn min(self, other: Self) -> Self

Get the min of self and other by element.

pub fn max(self, other: Self) -> Self

Get the max of self and other by element.

pub fn clamp(self, min: Self, max: Self) -> Self

Clamp self to the range [min, max] by element.

pub fn floor(self) -> Self

Floor the vector by element.

pub fn ceil(self) -> Self

Ceil the vector by element.

pub fn round(self) -> Self

Round the vector by element.

pub fn signum(self) -> Self

Get the absolute value of the vector.

pub fn length_squared(self) -> f32

Get the length of the vector squared.

pub fn length(self) -> f32

Get the length of the vector.

pub fn normalize(self) -> Self

Normalize the vector.

If the length of the vector is zero, the zero vector is returned.

pub fn dot(self, other: Self) -> f32

Get the dot product of self and other.

pub fn cross(self, other: Self) -> f32

Get the length of the cross product of self and other.

pub fn hat(self) -> Self

Hat the vector.

pub fn angle_between(self, other: Self) -> f32

Get the angle between self and other.

pub const fn to_point(self) -> Point

Convert the vector to a vector.

pub const fn to_size(self) -> Size

Convert the vector to a size.